finish part 2
this was fun. more parallelism to go fast.

main: kick off a goroutine for each row that calls divieRows. increment
our waitgroup. kick off another goroutine to watch the waitgroup and close
our result channel once we have all our results. pull from the prods
channel, adding each integer to sum as it arrives.

fan out: for each row, make a new channel to wait for the single result we
expect. for each element in the row, pass of the copy of that row with
the element removed along with our channel and the element itself to a
goroutine that will check if that element is divisible by any of the
memebrs of the new row. if it is, push the result of that division to the
channel and close the channel. one we receive the single result from the
channel, pass it along to the main's result channel.